pyproject.toml
[tool.poetry]
name = "platerec"
version = "0.1.0"
description = ""
authors = ["Marc Belmont"]

[tool.poetry.dependencies]
python = "^3.6.1"
Pillow = "^7.1"
requests = "^2.23.0"
opencv-python = "^4.2.0"
psutil = "^5.7.0"
streamlit = "^0.57.3"

[tool.poetry.dev-dependencies]
pre-commit = "^2.2.0"
pytest = "^5.4"
hypothesis = "*"
flake8 = "*"
yapf = "==0.27"
rope = "^0.14.0"
pdbpp = "*"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
